codegen: Move some stuff into a utils directory
